Add system to define Talon lists in CSV files
The CSVs are stored in a "settings" folder that the user can track
independently. When the CSVs are updated, the talon lists they map to
will be updated (in the `user.user_settings` context, which is global).

Both mapped items (key -> value) and atomic items (key -> key) can be
defined.