Fix editor commands being sent when popup windows are open
As related to this issue
https://github.com/anonfunc/intellij-voicecode/issues/7
when a non-editor window of a Jetbrains product is open we still
mark Jetbrains commands as active, meaning things like copy/paste
or 'clear line' go to the editor, rather than the popup window.

This was previously implemented using callbacks for the context in
the original Talon.
https://github.com/anonfunc/talon-user/blob/master/apps/jetbrains.py#L303

These have been replaced with scopes. The current scope updates when
the UI fires an event. This does mean if you open the 'actions' dialogue
(via the command please) then it doesn't trigger a UI update and the Jetbrains
context is still active. Other popups do. I'm not sure what the performance hit
would be running the check every second or two via cron so opted for this
instead