Fix watchdogs when getting page address in Safari
This sometimes inadvertently scanned in the web page. Now we only look in the window's toolbar for the address and search field.

Tested in all toolbar configurations in macOS 10.15 and 12.

Also add a short timeout to attempt to prevent hangs while trying to retrieve the frontmost tab's URL via AppleScript (backup method).