Links to the Talon and Knausj issue trackers
Added two commands that print out links to the Talon and Knausj
issue trackers to encourage bug reporting.  Right now the link to
the Talon issue tracker is hidden as a pinned message on the help
channel of the slack.  Attatching it to the command 'report Talon bug'
may make it easier to find.