Bugfixes for c, rust (#961)
Apologies, I accidentally closed the first pull request for this.

When these insert library functions were called in their respective
talon files, the first argument was the name of the library, and the
second argument was an empty string. As a result, it was pasting an
empty string ("selection") instead of the library name ("text").

Also, one of the capture rules for c didn't match the name used for it
in the associated talon file, causing the voice commands to not work.