edit_mac.py: omit cmd-left for file_start/file_end (#1057)
The default bindings for `edit.file_start` on mac presses `cmd-up
cmd-left` (and `edit.file_end` is `cmd-down cmd-left`). I find this
`cmd-left` to be redundant (at least in TextEdit and in Emacs, `cmd-up`
goes to the start of the document, including the beginning of the line)
and to produce bad behavior in, eg, web browsers, where `cmd-up`
successfully scrolls to the top but then `cmd-left` goes back in
history.