Refactor browser files, improve consistency and add functionality (#1117)
The goal of this PR is to reduce duplication, add consistency and
complete functionality where possible. It should close issue #532.

Things done:

- Move most of the common browser functionality to browser.py and
browser_mac.py
- Implement tab actions in those same two files and in the specific
browser apps when they differ.
- Change mac shortcuts for tab next and tab previous to `ctrl-tab` and
`ctrl-shift-tab` that have less issues with international keyboards.
(Maybe a bit out of the scope of this PR)
- Implement action `browser.focus_page` and add command to trigger it.
- Implement action `user.tab_duplicate` for all the browsers.
- Implement commands `bookmarks sidebar` and `history sidebar` and their
respective actions for Firefox.
- Fix some shortcuts that were incorrect.
- Raise NotImplementedError for the actions that couldn't be implemented
with a clear message of the reason.
- Add support for Opera and Vivaldi browsers.
- Add .talon files for all the browsers. This makes it easier if the
user wants to add custom commands.
- Move all the tag activation to .talon files
- Remove commented out commands in browser.talon. Many of those can now
be performed with Rango.