feat(#755) Pop to click setting for all users (#1139)
This PR addresses #755 and replaces #858 . It does two things:

1. The pop noise triggers an action `user.noise_trigger_pop`.
2. The existing setting `user.mouse_enable_pop_click` gets a third value
'2' which says always pop to click.

The first point lets people 'take over' the pop noise so they don't have
to worry about whatever knausj talon is doing with it. It also provides
a nice scaffold in case you want it to do different things in different
contexts.

The second point differs a bit from the pop_click = yes/no boolean
discussed in #755 . This is for backwards compatibility. With a yes/no
setting I'd either have to turn it on by default for people without eye
trackers (i.e. undo https://github.com/knausj85/knausj_talon/pull/1119)
or turn it off and break the expected behaviour for most eyetracker
users. If I figured out a way to conditionally enable the setting when
an eye tracker was on (not trivial given how you have to check for this)
then this would break for eyetracker users who had turned the setting
off because they didn't want pop to click.

Hopefully the addition of a new value is simple enough for people to use
if they want pop to click in a non-eyetracker context.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jeff Knaus <knaus.jeff@gmail.com>