Switch screenshot_rect to become a user action (#1160)
As part of writing a screenshot tool a long time ago, I changed
`screenshot_rect` to a user action. As I'm hoping to eventually push or
share the other tool for others to use, I figured I'd push this change.

For reference the other tool is called shotbox
https://github.com/fidgetingbits/knausj_talon/tree/master/plugin/shotbox,
and the usage of this action is specifically
https://github.com/fidgetingbits/knausj_talon/blob/master/plugin/shotbox/shotbox.py#L659

I vaguely recall when I wrote this the general consensus was that we
shouldn't be importing other local python files (I think because of
potential negative affects on talon dynamic module reloading), which is
why I did this as a user action instead of just `from .. import
screenshot_rect`. If this is no longer the case than it may be not
necessary and I can just change my shotbox code, so I'll leave it up to
you folks to decide.