Hide cmd window for mintty when setting cyg_path.
The cmd.exe window that the subprocess module uses does not start
hidden, and so steals focus from the mintty window.  This focus stealing
may happen several times.

The stolen focus is not restored unless Talon sets the focus via the
focus command, and inhibits usage of physical input devices until that
focus is restored.