jetbrains: reduce log output, better error output
until now, finding the vcidea_foo file still resulted in an error that
the .vcidea_foo file was not found, which is confusing. also, there's
always many lines per command executed, with this patch it's just
one when it works.
If there's an error with the vcidea nonce file, or if the port isn't found,
there will now be a hint that vscode-idea commands will not work,
and when the vcidea file is not found, the actual locations we looked
will be put into the log message as well, so you don't have to guess
or investigate with the REPL or something