Implemented compound edit command (#1498)
Compound of action(select, clear, copy, cut, paste, etc.) and
modifier(word, line, etc.) commands for editing text. eg: `"select
line"`, `"clear all"`

Fixes #1412
Fixes #1493

Note that there are support for some actions in here that I have not
added a spoken form for. "insert" for example.

https://github.com/talonhub/community/blob/df1dc66815c875bcc942649c65cf242da0231a51/core/edit/edit_command_actions.py#L75-L77

This one is a provision for Cursorless or something similar to use and
an example of how you can extend this implementation with more complex
actions compared to what you can do with just an identifier in a Talon
list. The other action `applyFormatter` is definitely something we can
wire up in a future update.



Also note that there are a lot of edit actions left out of this
implementation. In some cases because they don't really fit the
action/modifier paradigm like `"go left"`. Calling the modifier
`character` as `left` isn't really correct. And in other cases we can
definitely do a follow up and fill out this implementation some more.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: David Vo <auscompgeek@users.noreply.github.com>