Don't use not mode (#1515)
Users should be allowed to make their own custom modes and not had any
existing commands interfere.

This is what a Talon wiki describes.
https://talon.wiki/customization/talon%20framework/modes/

Fixes #1437



We could of course go even more explicit and not even have these
combination files. The sleep command could be in a command mode file and
in the dictation mode file etc.

Note that the dragon files are still enabled in all nodes. I don't know
this engine well enough to make changes and I don't want to break
something.

---------

Co-authored-by: Nicholas Riley <com-github@sabi.net>