Enable readline and python tags in talon repl (#1540)
Similar to #1451 this uses a context using not forced language to
override `code.language` and enable python lang features inside the
talon repl. It also makes the title matching more restricting, as this
currently matches on anything with repl in the name, and there are many
repls that aren't talon or python...

I also enable readline so that common navigation commands work inside.