Adds additional helpers for refactoring to intellij (#1425)
this pull request adds commands to go to a line and search for some text
and immediately perform an action on that. it makes it very simple to
operate on words that you can see without having access to cursorless.
This is very helpful in intellij since you don't have cursorless. It
also expands on existing refactoring commands by adding specific ones
for things like renaming and quick fixing.

---------

Co-authored-by: Nicholas Riley <com-github@sabi.net>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>