Extending the pages tag (#1455)
This pull request aims to extend the pages tag and applies the new
commands for foxit, Sumatra, and the adobe pdf reader.

Currently, the pages tag seems to be used by document viewers, mostly
pdf readers. This type of software typically has the ability to rotate
pages, and also to go back and forth when users click on an internal
link. this pull request adds these functions to the pages tag.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>
Co-authored-by: Jeff Knaus <knaus.jeff@gmail.com>