Move command client into core folder (#1678)
The reason the command client resides in the vscode folder is because it
was designed for vscode originally, but now it's actually used for more
applications so I'm moving it to the core folder instead.

I will take care of the subtree. I have successfully run the split
command on my computer.