Make the word wheel optional for stop continuous scrolling commands (#1749)
Allows stopping continuous scrolling with "stop" and "stop here". 

Something worth discussing: should we indicate the grammar change in the
continuous scrolling graphical display and if so how? Should we just
replace "wheel stop" in the display with "stop" or should we indicate
that the word wheel can optionally be used?

---------

Co-authored-by: Nicholas Riley <com-github@sabi.net>