Remove /settings from default .gitignore.
We have migrated most of the files that used to be in this directory to be colocated with the .py/.talon implementations of the corresponding features, therefore this is a move to consistency with our policy (in the README) that you should create a fork of community to house your changes and customizations.

Fixes #1732.