Start migrating elixir to snippets (#1880)
There does not seem to be an else if statement in the language, so I did
not try to migrate that action.
I was not sure what to do about the fact that a switch statement is
called a case statement in the language, so changing the code state case
action to do what that action usually does would not reflect what things
are named in this language.
There is an issue with actions conflicting with commands with the elixir
setup, so I migrated the actions but left the commands alone.
I followed the structure in the existing talon file of explicitly
defining state commands.
I am not familiar enough with the language to know how to properly
handle the import actions, which are currently unusable anyways because
they do not correspond to action definitions.

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>