Start migrating go to snippets (#1862)
I was admittedly not sure what to do with the for loop action. I will
leave it to someone more familiar with the language to figure that out.
The for loop and while loop actions are currently equivalent because the
for key word is used in go for while loops.

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>