Migrate lua try catch (#1899)
I put the snippet in the lua file because the try catch statement file
requires 2 snippet placeholders. I was not sure how to properly convert
this into a snippet due to lack of familiarity with the language, so I
just recreated the functionality of the original action and will leave a
proper snippet definition to someone more familiar with the language.

Co-authored-by: Nicholas Riley <com-github@sabi.net>