Have raw snippet insert reach the first stop from the end of the line (#1909)
The raw snippet insert currently reaches the first stop by moving to the
correct column, going to the beginning of the line, and pressing right.
This can fail when there is indentation at the start of the line by
putting the cursor in the wrong position. This pull request addresses
this by moving the cursor to the end of the line and then moving left
the required number of times.