Address c signed spacing (#1913)
I attempt to address at least part of this issue:
https://github.com/talonhub/community/issues/853. The c_signed list had
the issue of having spaces at the end of its words, which would create
an extra space when used by captures not expecting that. I addressed
this by removing the extraneous spaces and adding a space at the end of
the command consisting solely of the c_signed capture.
I have not done c programming in ages, so I would appreciate it if
someone actually using this grammar could test to see if this breaks
anything before it gets merged.