Add setting for determining context sensitive dictation peek character (#1915)
In response to this issue:
https://github.com/talonhub/community/issues/1817, this pull request
adds a setting for determining the character inserted during context
sensitive dictation without application support to handle the case of
nothing before or after the cursor. This would make changing the
character on a per application basis easier and let someone pick a
better default for their use case.

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>