Include wakeup commands in command history (#1916)
Fixes: https://github.com/talonhub/community/issues/793.

Including wakeup commands in the history it is useful if a user woke up
talon without knowing what command did that and wants to know what
happened. This also improves consistency between the command history and
the talon subtitles.

I did a little refactoring to simplify the code and add more type
annotations to the variables.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>