Add clipboard help search command (#1902)
Do we want to address this issue
https://github.com/talonhub/community/issues/1562? If so, here is an
implementation. Other commands, in particular the screenshot ones, use
clip to refer to the clipboard, so I used that here. I put the word clip
before the word search in case someone wanted to do a help search on the
word clip itself.

---------

Co-authored-by: Nicholas Riley <com-github@sabi.net>