Add snippet stop at the end (#1934)
Add a snippet stop at the end of every snippet that does not already
have one there to allow exiting the snippet with "snip next".

I replaced .* with .*? in the regular expression to match the smallest
possible match to better handle multiple instances of the same snippet
stop with default values.

Someone should probably test this other than me as breaking snippets
would not be fun.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nicholas Riley <com-github@sabi.net>