Update formatters.py
This change adds functions to get the prose and word formatter specifically so that in the next change we can make help commands to get those.

I have found personally that what I am trying to think about the formatters the help menu that has the code and the prose together is too confusing. I like being able to specifically see the prose or word formatters when I am trying to learn or recall specific commands.