Mouse wake enables controls mouse and no zoom
This behavior was extremely confusing, we're telling talon to `sleep all` would shut everything off but having it wake up from that with `welcome_back` would only activate zoom mode on the mouses. It took me several minutes to realize this unusual behavior.

The function descriptor was totally incorrect because it did not actually enable controlling the mouse, it only enabled the zoom.

I think most users are going to want to default to waking up to having control over the mouse and not having zoom mode enabled; while zoom mode is useful, I think most users would use the control setting most of the time.