Expand formatter help commands
I often find myself trying to see just the words for a particular type
of formatter. This is clearly missing in terms of the word formatters,
and while it does not matter significantly in terms of the prose and
code formatters it is still nice to be able to see a separation there.