[Quick Fix] Do temporary fix given Snippet replacement issue (#2001)
The decision to insert $0 at the end of snippets has messed with snippet
replacements involving $0. This should fix all of them.
For a longer term fix, I recommend we stop changing the snippet when
loading it and instead have an action that changes the snippet during
insertion after making replacements.