cpp partial object oriented support (#2014)
This is a little complicated by accessing self in C++ being `this->` or
`*(this).` instead of `this.`, but the standard command is `self dot`. I
made questionable choices to accommodate the existing spoken form and
the different things you do with `this` in C++.

I made `self dot` do `*(this).` to be consistent with the name of the
command while actually accessing the object properly. I made `self dot`
now use an overwritable action that defaults to the prior behavior.

I added custom actions for C++ based on common ways to interact with the
`this` keyword.

```talon
self arrow: insert("this->")
(star | dereference) self: insert("*this")
```

@chdoc do you think this is a reasonable way to handle this? Do you have
alternative suggestions?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>