fix waterfox support on macOS (#2080)
What it says on the tin. Very simple fix, just requires adding the app
bundle information to `firefox.py`. Also added an alternate
capitalization to the 'simple' app identifiers just in case.

If someone who daily drives Windows and Linux can double-check both of
those are working as well that'd be good; any fixes needed there could
be added to this as well (but I think the previous Waterfox pr should've
worked fine for both of those?).

Co-authored-by: Nicholas Riley <com-github@sabi.net>